requirements.txt
DateTime==4.5
display==1.0.0
Flask==2.1.2
Flask-SQLAlchemy==2.5.1
flatbuffers==1.12
fonttools==4.36.0
gast==0.4.0
gitdb==4.0.9
GitPython==3.1.27
gunicorn==20.1.0
httpcore==0.15.0
httpx==0.23.0
hyperframe==5.2.0
Jinja2==3.1.2
jinja2-time==0.2.0
jsonschema==4.5.1
keras==2.9.0
lightgbm==3.3.2
Markdown==3.4.1
matplotlib==3.5.3
matplotlib-inline==0.1.3
numpy==1.22.3
packaging==21.3
pandas==1.4.2
python-dateutil==2.8.2
python-dotenv==0.20.0
python-slugify==6.1.2
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
pyzmq==22.3.0
Quandl==3.7.0
regex==2022.8.17
requests==2.27.1
requests-oauthlib==1.3.1
rfc3986==1.5.0
rsa==4.8
sample-data==0.1.0
scikit-learn==1.1.1
scipy==1.8.1
shellescape==3.8.1
six==1.16.0
SQLAlchemy==1.4.37
stack-data==0.2.0
tensorflow-cpu==2.9.1
text-unidecode==1.3
threadpoolctl==3.1.0
tinycss2==1.1.1
tokenization==1.0.7
toml==0.10.2
tqdm==4.64.0
urllib3==1.26.9
